pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "setuptools-scm>=8.0",
]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
]
name = "nomad-material-processing"
dynamic = ["version"]
description = "A plugin for NOMAD containing base sections for material processing."
readme = "README.md"
authors = [
    { name = "Andrea Albino" },
    { name = "Sebastian Brückner" },
    { name = "Michael Götte" },
    { name = "Ahmed Ilyas" },
    { name = "Sarthak Kapoor" },
    { name = "José A. Márquez" },
    { name = "Hampus Näsström" },
    { name = "Markus Scheidgen" },
    { email = "fairmat@physik.hu-berlin.de" },
]
maintainers = [
    { name = "FAIRmat", email = "fairmat@physik.hu-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab >= 1.3.14",
]

[project.urls]
"Homepage" = "https://github.com/FAIRmat-NFDI/nomad-material-processing"
"Bug Tracker" = "https://github.com/FAIRmat-NFDI/nomad-material-processing/issues"

[tool.uv]
index-url = "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple"


[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "structlog",
    "mkdocs", 
    "mkdocs-material>=9.0",
    "pymdown-extensions",
    "mkdocs-click",
    "mkdocs-macros-plugin>=1.0",
    "pydantic>=2.0,<2.11",
    "openpyxl",
]

[tool.ruff]
include = ["src/*.py", "tests/*.py"]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']

general_schema = "nomad_material_processing:schema"
solution_schema = "nomad_material_processing.solution:schema"
vd_schema = "nomad_material_processing.vapor_deposition:schema"
cvd_schema = "nomad_material_processing.vapor_deposition.cvd:schema"
movpe_schema = "nomad_material_processing.vapor_deposition.cvd:movpe_schema"
pvd_schema = "nomad_material_processing.vapor_deposition.pvd:schema"
mbe_schema = "nomad_material_processing.vapor_deposition.pvd:mbe_schema"
pld_schema = "nomad_material_processing.vapor_deposition.pvd:pld_schema"
sputtering_schema = "nomad_material_processing.vapor_deposition.pvd:sputtering_schema"
thermal_schema = "nomad_material_processing.vapor_deposition.pvd:thermal_schema"
sintering_schema = "nomad_material_processing.sintering:schema"
measurement_schema = "nomad_material_processing.measurements:schema"

[tool.setuptools_scm]
